add python code to run docker unittests
The tests are meant to be run via test-ci.sh.
This script functions similarly to build-ci.sh and deploy-sh.
It requires $DOCKERHUB_ORG to be set and uses git to deterimine
files that are changed. If either Dockerfile or unitest.yml files
have been changed the tests in the unittest.yml file will be run.